moves grid and list view tabs into main view
